Refactored Functional tests into one extendable test for all fixtures.
