--- tests/Functional/BeaumontConsultingTest.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Tests\Functional;
-
-use Tests\Fixtures\LoadsFixtures;
-use Tests\TestCase;
-
-class BeaumontConsultingTest extends TestCase
-{
-    use LoadsFixtures, Purifiable;
-}
