--- tests/Functional/MissionAustraliaTest.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Tests\Functional;
-
-use Tests\Fixtures\LoadsFixtures;
-use Tests\TestCase;
-
-class MissionAustraliaTest extends TestCase
-{
-    use LoadsFixtures, Purifiable;
-}
