--- tests/Functional/RedCrossAusTest.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Tests\Functional;
-
-use Tests\Fixtures\LoadsFixtures;
-use Tests\TestCase;
-
-class RedCrossAusTest extends TestCase
-{
-    use LoadsFixtures, Purifiable;
-}
